requirements.txt
albumentations==0.4.6
einops==0.4.1
torchvision
wandb
segmentation-models-pytorch

